API: libpe_status,libcrmcommon: move enum pe_obj_types to libcrmcommon
